+++ CHANGELOG.md
@@ -8,8 +8,8 @@
   - `mustBeNonempty`, `mustBeNonNan`, `mustBeNonZero`, `mustBeNonSparse`,
   - `mustBeA`, `mustBeReal`, `mustBeInteger`, `mustBeNonmissing`,
   - `mustBePositive`, `mustBeNonpositive`, `mustBeNonnegative`, `mustBeNegative`,
-  - `mustBeGreaterThan`, `mustBeGreaterThanOrEqual`, `mustBeLessThan`.
-  - `mustBeNumericOrLogical`
+  - `mustBeGreaterThan`, `mustBeGreaterThanOrEqual`, `mustBeLessThan`,
+  - `mustBeNumericOrLogical`, `mustBeText`.
 
 - `test_run` manages `SEQUENTIAL TEST REQUIRED` and `NATIVE_ARCHITECTURE TEST REQUIRED` tags.